+++ flika/flika.py
@@ -6,7 +6,7 @@ import sys, os
 import optparse
 
 # for development purposes, add this if flika is not in your site-packages
-#sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
+sys.path.insert(0, os.path.dirname(os.path.dirname(__file__)))
 
 #from flika import __version__
 from flika import __version__
